pyproject.toml
[project]
name = "api-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "boto3>=1.41.0",
    "docling[rapidocr,vlm]>=2.62.0",
    "fastapi[standard]>=0.121.2",
    "aiosmtplib>=3.0.1",
    "jinja2>=3.1.4",
    "google-genai>=1.51.0",
    "openai>=2.8.1",
    "psycopg[binary,pool]>=3.2.12",
    "pydantic-settings>=2.12.0",
    "pymilvus[milvus-lite]>=2.6.3",
    "python-jose>=3.5.0",
    "sentence-transformers>=5.1.2",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "greenlet>=3.2.4",
    "langchain-docling>=2.0.0",
    "langchain-core>=1.1.0",
    "langchain-huggingface>=1.0.1",
    "langchain-milvus>=0.3.0",
    "langchain>=1.0.8",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "matplotlib>=3.10.7",
    "langchain-openai>=1.0.3",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
